Added JWT&OAuth2 authenticate

+++ routes/events.py
@@ -1,8 +1,9 @@
-from fastapi import APIRouter, Body, HTTPException, status
+from fastapi import APIRouter, Body, HTTPException, status, Depends
 from beanie import PydanticObjectId
 from database.connection import Database
 from models.events import Event, EventUpdate
 from typing import List
+from auth.authenticate import authenticate
 
 event_router = APIRouter(
     tags=["Events"]
@@ -30,7 +31,8 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
 
 
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         "message": "Event created successfully."
@@ -38,7 +40,13 @@ async def create_event(body: Event) -> dict:
 
 
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
@@ -49,7 +57,13 @@ async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
 
 
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     event = await event_database.delete(id)
     if not event:
         raise HTTPException(
@@ -62,7 +76,7 @@ async def delete_event(id: PydanticObjectId) -> dict:
 
 
 @event_router.delete("/")
-async def delete_all_events() -> dict:
+async def delete_all_events(usr: str = Depends(authenticate)) -> dict:
     await event_database.delete_all()
     return {
         "message": "Events deleted succesfully."
